unsigned to size_t conversion

+++ tensor_compare_lib.h
@@ -5,10 +5,10 @@
 namespace tensor_lib {
 
 template <typename T>
-bool compareArray(const T* A, const T* B, const unsigned Size);
+bool compareArray(const T* A, const T* B, const size_t Size);
 
 template <typename T>
-bool compare2Darray(const T** A, const T** B, const unsigned SizeY, const unsigned SizeX);
+bool compare2Darray(const T** A, const T** B, const size_t SizeY, const size_t SizeX);
 
 template <typename T>
 bool compareTensors(const std::vector<T> & A, const std::vector<T> & B);
@@ -24,10 +24,10 @@ bool compareTensors(const std::vector<std::vector<std::vector<std::vector<T> > >
 
 // W. Threshold
 template <typename T>
-bool compareArray(const T* A, const T* B, const unsigned Size, const T t);
+bool compareArray(const T* A, const T* B, const size_t Size, const T t);
 
 template <typename T>
-bool compare2Darray(const T** A, T** B, const unsigned SizeY, const unsigned SizeX, const T t);
+bool compare2Darray(const T** A, T** B, const size_t SizeY, const size_t SizeX, const T t);
 
 template <typename T>
 bool compareTensors(const std::vector<T> & A, const std::vector<T> & B, const T t);
